Show remote files in linear list
